// apps/api/src/posts/posts.service.ts
export const postInclude = {
  route: { include: { source: true, target: true } },
  media: { orderBy: { messageId: 'asc' as const } },
  revisions: { orderBy: { version: 'desc' as const } },
};